# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartTrainingRequest": {
        "properties": {
          "base_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Model"
          },
          "batch_size": {
            "default": 8,
            "title": "Batch Size",
            "type": "integer"
          },
          "epochs": {
            "default": 3,
            "title": "Epochs",
            "type": "integer"
          },
          "extra_jsonl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Jsonl"
          },
          "lora": {
            "default": false,
            "title": "Lora",
            "type": "boolean"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "seed": {
            "default": 17,
            "title": "Seed",
            "type": "integer"
          },
          "val_ratio": {
            "default": 0.1,
            "title": "Val Ratio",
            "type": "number"
          }
        },
        "required": [
          "model"
        ],
        "title": "StartTrainingRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "**한국어 사기 신호 검출 reference implementation** — 음성·텍스트·이미지·PDF·URL 입력에서 위험 신호를 검출하고 각 신호의 학술/법적 근거를 transparent 하게 보고한다.\n\n⚠️ **Identity Boundary** (CLAUDE.md): ScamGuardian 은 *사기 판정을 내리지 않는다*. VirusTotal 이 70개 백신의 검출 결과를 보고만 하는 모델과 동일. 판정 logic 은 통합한 기업(통신사·은행·메신저 앱)이 자기 risk tolerance 에 따라 구현한다.\n\n- **Public 4 endpoint** (`Public` 태그) — 외부 통합 시작점, `DetectionReport` 응답\n- **Webhook** (`Webhook` 태그) — 카카오 오픈빌더\n- **Admin** (3 태그) — 라벨링 / platform / training\n- **v4 draft** — Live Call Guard 설계 preview (구현 X)\n\n통합 가이드: [`docs/INTEGRATION_GUIDE.md`](https://github.com/example/scamguardian-v2/blob/main/docs/INTEGRATION_GUIDE.md)",
    "title": "ScamGuardian Signal Detection API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/training/data-stats": {
      "get": {
        "description": "현재 라벨링 데이터 라벨 분포 + 엔티티 수 — 학습 시작 전 충분성 판단용.",
        "operationId": "admin_training_data_stats_api_admin_training_data_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Training Data Stats Api Admin Training Data Stats Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "학습 데이터 통계",
        "tags": [
          "Admin — Training"
        ]
      }
    },
    "/api/admin/training/sessions": {
      "get": {
        "description": "모든 세션 메타 + 현재 파이프라인이 사용하는 active 모델 경로 (`active_models.json`).",
        "operationId": "admin_training_list_api_admin_training_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Training List Api Admin Training Sessions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "학습 세션 목록 + 활성 모델",
        "tags": [
          "Admin — Training"
        ]
      },
      "post": {
        "description": "subprocess 로 학습 세션 spawn — `.scamguardian/training_sessions/{id}/` 에 `status.json` / `metrics.jsonl` / `train.log` 출력.\n\n**Body** (`StartTrainingRequest`):\n- `model` — `classifier` (mDeBERTa) 또는 `gliner`\n- `epochs` (기본 3), `batch_size` (기본 8), `lora` (LoRA 사용)\n- `extra_jsonl` — 추가 데이터셋 경로\n- `val_ratio` (기본 0.1), `seed` (기본 17), `base_model`",
        "operationId": "admin_training_start_api_admin_training_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Training Start Api Admin Training Sessions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "유효성 실패"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "fine-tune 세션 시작",
        "tags": [
          "Admin — Training"
        ]
      }
    },
    "/api/admin/training/sessions/{session_id}": {
      "get": {
        "description": "`session` 메타 + 마지막 500 metric 이벤트 + 마지막 8KB 로그.",
        "operationId": "admin_training_detail_api_admin_training_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Training Detail Api Admin Training Sessions  Session Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "404": {
            "description": "session not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "세션 상세 + metrics tail + log tail",
        "tags": [
          "Admin — Training"
        ]
      }
    },
    "/api/admin/training/sessions/{session_id}/activate": {
      "post": {
        "description": "체크포인트 경로를 `.scamguardian/active_models.json` 에 기록 → `pipeline.active_models` 60s TTL 캐시가 무효화되어 즉시 swap.\n\n분류기 / GLiNER 각 1개씩 활성 가능. 경로 무효 시 base 모델로 fallback.",
        "operationId": "admin_training_activate_api_admin_training_sessions__session_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Training Activate Api Admin Training Sessions  Session Id  Activate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "유효성 실패 (e.g. 체크포인트 경로 없음)"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "404": {
            "description": "session not found 또는 모델 파일 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "학습 결과를 파이프라인에 적용",
        "tags": [
          "Admin — Training"
        ]
      }
    },
    "/api/admin/training/sessions/{session_id}/cancel": {
      "post": {
        "description": "실행 중 subprocess 종료 + status `cancelled` 갱신.",
        "operationId": "admin_training_cancel_api_admin_training_sessions__session_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Training Cancel Api Admin Training Sessions  Session Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "어드민 토큰 누락 또는 무효"
          },
          "409": {
            "description": "취소할 수 없는 상태 (이미 끝남)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        },
        "summary": "학습 세션 취소",
        "tags": [
          "Admin — Training"
        ]
      }
    },
...
  },
...
}